requirements.txt
ansible==9.3.0
ansible-core==2.16.4
cffi==1.16.0
cryptography==42.0.5
Jinja2==3.1.3
MarkupSafe==2.1.5
networkx==3.2.1
packaging==23.2
pycparser==2.21
PyYAML==6.0.1
resolvelib==1.0.1
